+++ app/src/main/java/nbc/group/recipes/data/model/dto/SpecialtyResponse.kt
@@ -1,5 +1,6 @@
 package nbc.group.recipes.data.model.dto
 
+import com.google.gson.annotations.SerializedName
 import com.tickaroo.tikxml.annotation.Element
 import com.tickaroo.tikxml.annotation.PropertyElement
 import com.tickaroo.tikxml.annotation.Xml
