--- app/src/main/java/nbc/group/recipes/presentation/fragment/home/HomeQuizAdapter.kt
@@ -1,51 +1,0 @@
-package nbc.group.recipes.presentation.fragment.home
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import nbc.group.recipes.databinding.ItemHomeQuizBinding
-import nbc.group.recipes.util.RecipeItem
-
-// RecipeItem 더미 데이터 테스트
-val diffCallbackHomeQuiz = object : DiffUtil.ItemCallback<RecipeItem>() {
-    override fun areItemsTheSame(oldItem: RecipeItem, newItem: RecipeItem): Boolean {
-        return oldItem.item == newItem.item
-    }
-
-    override fun areContentsTheSame(oldItem: RecipeItem, newItem: RecipeItem): Boolean {
-        return oldItem == newItem
-    }
-}
-
-class HomeQuizAdapter :
-    ListAdapter<RecipeItem, HomeQuizAdapter.ItemViewHolder>(diffCallbackHomeQuiz) {
-
-    // 이미지 데이터 소스?
-    // 퀴즈 로직?
-
-    class ItemViewHolder(
-        private val binding: ItemHomeQuizBinding,
-    ) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(items: RecipeItem) {
-            binding.apply {
-                Glide.with(root.context)
-                    .load(items.imageResourceId)
-                    .into(ivQuiz)
-            }
-
-        }
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ItemViewHolder {
-        val binding =
-            ItemHomeQuizBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return ItemViewHolder(binding)
-    }
-
-    override fun onBindViewHolder(holder: ItemViewHolder, position: Int) {
-        holder.bind(getItem(position))
-    }
-}
