--- app/src/main/java/nbc/group/recipes/util/Specialties.kt
@@ -1,53 +1,0 @@
-package nbc.group.recipes.util
-
-// 특산품 종류별로 API 데이터를 받아오기 위한 변수 선언
-
-// 곡물류
-val specialties1 = listOf(
-    "잡곡", "밀가루", "옥수수", "율무"
-)
-
-// 채소류
-val specialties2 = listOf(
-    "취나물", "미나리", "잎들깨", "오이", "부추", "깻잎", "대파", "배추", "미나리", "고추", "쑥", "상추", "고사리",
-    "파프리카", "시래기", "가지", "쪽파", "명이", "부지갱이", "미역취", "서덜취", "머위", "음나무", "차나무", "여주",
-    "양상추", "시금치", "박나물", "치자", "알로에", "브로콜리", "단호박", "호박", "애호박", "쥬키니", "고구마줄기",
-    "고구마", "감자", "당근", "연근", "마늘", "우엉", "양파", "무", "도라지", "순무", "돼지감자", "더덕", "생강",
-    "마", "천마", "토란", "인삼", "산삼", "산양삼", "눈개승마", "어성초", "피망", "곤드레"
-    )
-
-// 과일류
-val specialties3 = listOf(
-    "참외", "멜론", "체리", "복숭아", "포도", "사과", "수박", "단감", "블루베리", "토마토", "딸기", "감", "배", "복숭아",
-    "다래", "구기자", "오미자", "대추", "복분자", "무화과", "한라봉", "유자", "대봉", "자두", "망고", "애플망고", "용과",
-    "키위", "매실", "산딸기", "밤", "오디", "두릅", "가시오가피"
-)
-
-// 어류
-val specialties4 = listOf(
-    "붕장어", "곰장어", "갈치", "멸치", "장어", "숭어", "황복", "참게", "코다리", "오징어", "전어", "은어", "철갑상어",
-    "쥐치", "황태", "어란"
-)
-
-// 해조류
-val specialties5 = listOf(
-    "다시마", "미역", "김"
-)
-
-// 고기류
-val specialties6 = listOf(
-    "한우", "닭", "오리", "양돈", "녹돈", "돼지", "유정란", "율란", "달걀", "닭", "오리", "노루궁뎅이"
-)
-
-// 쌀
-val specialties7 = listOf(
-    "쌀", "찹쌀", "흑미", "현미"
-)
-
-// 기타
-val specialties8 = listOf(
-    "참게", "코다리", "오징어", "전어", "은어", "철갑상어", "쥐치", "조개", "가리비", "전복", "새조개", "바지락",
-    "송이", "표고", "느타리", "영지", "상황", "양송이", "팽이버섯", "새송이버섯", "굴", "개조개", "문어", "대게",
-    "곤충", "꿀", "콩", "들깨", "호두", "잣", "녹차", "곰취", "삼백초",
-    // "기타 전부"
-)
