+++ app/src/main/java/nbc/group/recipes/viewmodel/MainViewModel.kt
@@ -5,7 +5,6 @@ import androidx.lifecycle.viewModelScope
 import com.google.firebase.auth.FirebaseUser
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import nbc.group.recipes.data.model.dto.Recipe
@@ -100,10 +99,11 @@ class MainViewModel @Inject constructor(
     private val _specialties = MutableStateFlow<SpecialtyResponse?>(null)
     val specialties = _specialties.asStateFlow()
 
-    fun doTest(specialtyType: String) {
+    fun doTest() {
         viewModelScope.launch {
-            val items = repository.getSpecialty("포항").body.items.item
-            _specialties.emit(items)
+            _specialties.emit(
+                repository.getSpecialty("양구")
+            )
         }
     }
 }
