--- app/src/main/java/nbc/group/recipes/viewmodel/SharedViewModel.kt
@@ -1,15 +1,0 @@
-package nbc.group.recipes.viewmodel
-
-import androidx.lifecycle.ViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import nbc.group.recipes.util.KindItem
-
-class SharedViewModel : ViewModel() {
-    private val _selectedKindItem = MutableStateFlow<KindItem?>(null)
-    val selectedKindItem = _selectedKindItem.asStateFlow()
-
-    fun setSelectedKindItem(kindItem: KindItem) {
-        _selectedKindItem.value = kindItem
-    }
-}
