--- app/src/main/res/drawable/search.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="512dp"
-    android:height="512dp"
-    android:viewportWidth="512"
-    android:viewportHeight="512">
-  <path
-      android:pathData="M193,1.1c-49.4,5.1 -95.1,26.9 -130.6,62.3 -24.6,24.6 -41.2,51.4 -51.8,83.6 -14.1,42.8 -14.1,87.2 0,130 10.6,32.3 27.3,59.2 52,83.7 48.8,48.5 115.3,70.5 183.9,60.8 35.3,-5 70.7,-19.5 98.5,-40.4 3,-2.3 5.7,-4 6.1,-3.9 0.4,0.2 30.7,29.8 67.3,65.8 41.7,41 68,66.1 70.4,67.2 9.8,4.6 20.8,-1.2 22.8,-12 1.6,-8.4 7.2,-2.4 -102.4,-110.6l-34.2,-33.8 5.8,-6.8c25,-28.9 41.5,-65.5 47.8,-105.5 2.4,-15.6 2.4,-44.2 0,-60 -7.1,-45.9 -27.2,-85.3 -60.7,-118.6 -46,-45.7 -110.1,-68.4 -174.9,-61.8zM240.9,34.6c40.4,5.9 75.9,23.5 104,51.5 34.4,34.5 52.3,77.4 52.3,125.9 0,17.9 -1.6,30.2 -5.8,45.7 -21.2,78.8 -93.1,133.3 -175.9,133.3 -49.3,-0 -95.1,-18.7 -129.3,-52.9 -23.2,-23.2 -38.7,-50 -47.1,-81.1 -7.4,-27.7 -7.4,-62.3 0,-90 16.8,-62.6 66.5,-111.9 129.2,-128.1 22.8,-5.9 50.7,-7.6 72.6,-4.3z"
-      android:fillColor="#000000"
-      android:strokeColor="#00000000"/>
-</vector>
