--- app/src/main/res/layout/fragment_specialty.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/specialtyFragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".presentation.fragment.specialty.SpecialtyFragment">
-
-    <!--    <androidx.cardview.widget.CardView-->
-    <!--        android:id="@+id/cvSpecialtyKind"-->
-    <!--        android:layout_width="150dp"-->
-    <!--        android:layout_height="150dp"-->
-    <!--        app:cardCornerRadius="180dp">-->
-
-    <!--        <ImageView-->
-    <!--            android:id="@+id/ivSpecialtyKind"-->
-    <!--            android:layout_width="150dp"-->
-    <!--            android:layout_height="150dp"-->
-    <!--            android:scaleType="centerCrop"-->
-    <!--            android:src="@drawable/img_specialty_sample7"-->
-    <!--            app:layout_constraintEnd_toEndOf="parent"-->
-    <!--            app:layout_constraintStart_toStartOf="parent"-->
-    <!--            app:layout_constraintTop_toTopOf="parent" />-->
-
-    <!--    </androidx.cardview.widget.CardView>-->
-
-    <TextView
-        android:id="@+id/tvSpecialtyKind"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="32dp"
-        android:layout_marginEnd="280dp"
-        android:text="곡물류"
-        android:textSize="30dp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toTopOf="@+id/recyclerContainer"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.core.widget.NestedScrollView
-        android:id="@+id/recyclerContainer"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_marginTop="@dimen/default_padding"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/tvSpecialtyKind">
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/recyclerViewSpecialtyDetail"
-            style="@style/content_text_style"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:padding="@dimen/default_padding"
-            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-            app:spanCount="2"
-            tools:listitem="@layout/item_specialty" />
-
-    </androidx.core.widget.NestedScrollView>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
