--- app/src/main/res/layout/item_home_kind.xml
@@ -1,41 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="95dp"
-    android:layout_height="wrap_content"
-    android:layout_marginStart="@dimen/default_padding"
-    android:layout_marginEnd="@dimen/top_banner_left_icon_margin"
-    android:layout_marginTop="10dp"
-    android:gravity="center"
-    android:orientation="vertical">
-
-    <FrameLayout
-        android:id="@+id/FrameLayout"
-        android:layout_width="95dp"
-        android:layout_height="95dp">
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            app:cardCornerRadius="48dp">
-
-            <ImageView
-                android:id="@+id/ivKind"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:scaleType="centerCrop"
-                android:src="@drawable/img_specialty_sample7" />
-        </androidx.cardview.widget.CardView>
-
-    </FrameLayout>
-
-    <TextView
-        android:id="@+id/tvKind"
-        style="@style/content_text_style"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="4dp"
-        android:maxLines="1"
-        android:gravity="center"
-        android:text="쌀"
-        android:textStyle="bold" />
-</LinearLayout>
