--- app/src/main/res/layout/item_home_quiz.xml
@@ -1,61 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="145dp"
-    android:layout_height="wrap_content"
-    android:layout_marginStart="@dimen/top_banner_left_icon_margin"
-    android:layout_marginEnd="@dimen/top_banner_left_icon_margin"
-    android:gravity="start"
-    android:orientation="vertical">
-
-    <LinearLayout
-        android:layout_width="160dp"
-        android:layout_height="150dp"
-        android:layout_marginTop="6dp"
-        android:orientation="vertical">
-
-        <FrameLayout
-            android:id="@+id/FrameLayout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <androidx.cardview.widget.CardView
-                android:layout_width="130dp"
-                android:layout_height="130dp"
-                android:layout_marginStart="2dp"
-                android:layout_marginEnd="2dp"
-                app:cardCornerRadius="15dp">
-
-                <ImageView
-                    android:id="@+id/ivQuiz"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:scaleType="centerCrop"
-                    android:src="@drawable/img_recipe_sample" />
-            </androidx.cardview.widget.CardView>
-
-        </FrameLayout>
-
-        <LinearLayout
-            android:layout_width="120dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:layout_marginEnd="8dp"
-            android:orientation="vertical">
-
-            <!--            <TextView-->
-            <!--                android:id="@+id/tvQuiz"-->
-            <!--                style="@style/content_text_style"-->
-            <!--                android:layout_width="match_parent"-->
-            <!--                android:layout_height="wrap_content"-->
-            <!--                android:layout_marginStart="2dp"-->
-            <!--                android:layout_marginTop="4dp"-->
-            <!--                android:ellipsize="end"-->
-            <!--                android:maxLines="1"-->
-            <!--                android:text="요리명"-->
-            <!--                android:textStyle="bold" />-->
-
-        </LinearLayout>
-
-    </LinearLayout>
-
-</LinearLayout>
